% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{fit_mle.Binomial}
\alias{fit_mle.Binomial}
\title{Fit a Binomial distribution to data}
\usage{
\method{fit_mle}{Binomial}(d, x, ...)
}
\arguments{
\item{d}{A `Binomial` object.}

\item{x}{A vector of zeroes and ones.}

\item{...}{Unused.}
}
\value{
a `Binomial` object
}
\description{
The fit distribution will inherit the same `size` parameter as
the `Binomial` object passed.
}
