% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{fit_mle.LogNormal}
\alias{fit_mle.LogNormal}
\title{Fit a Log Normal distribution to data}
\usage{
\method{fit_mle}{LogNormal}(d, x, ...)
}
\arguments{
\item{d}{A `LogNormal` object created by a call to [LogNormal()].}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A `LogNormal` object.
}
\description{
Fit a Log Normal distribution to data
}
\seealso{
Other LogNormal distribution: 
\code{\link{cdf.LogNormal}()},
\code{\link{pdf.LogNormal}()},
\code{\link{quantile.LogNormal}()},
\code{\link{random.LogNormal}()}
}
\concept{LogNormal distribution}
