% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frechet.R
\name{quantile.Frechet}
\alias{quantile.Frechet}
\title{Determine quantiles of a Frechet distribution}
\usage{
\method{quantile}{Frechet}(x, probs, drop = TRUE, elementwise = NULL, ...)
}
\arguments{
\item{x}{A `Frechet` object created by a call to [Frechet()].}

\item{probs}{A vector of probabilities.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{elementwise}{logical. Should each distribution in \code{x} be evaluated
at all elements of \code{probs} (\code{elementwise = FALSE}, yielding a matrix)?
Or, if \code{x} and \code{probs} have the same length, should the evaluation be
done element by element (\code{elementwise = TRUE}, yielding a vector)? The
default of \code{NULL} means that \code{elementwise = TRUE} is used if the
lengths match and otherwise \code{elementwise = FALSE} is used.}

\item{...}{Arguments to be passed to \code{\link[revdbayes]{qgev}}.
Unevaluated arguments will generate a warning to catch mispellings or other
possible errors.}
}
\value{
In case of a single distribution object, either a numeric
  vector of length `probs` (if `drop = TRUE`, default) or a `matrix` with
  `length(probs)` columns (if `drop = FALSE`). In case of a vectorized
  distribution object, a matrix with `length(probs)` columns containing all
  possible combinations.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

set.seed(27)

X <- Frechet(0, 2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
