% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReversedWeibull.R
\name{RevWeibull}
\alias{RevWeibull}
\title{Create a reversed Weibull distribution}
\usage{
RevWeibull(location = 0, scale = 1, shape = 1)
}
\arguments{
\item{location}{The location (maximum) parameter \eqn{m}.
`location` can be any real number.  Defaults to `0`.}

\item{scale}{The scale parameter \eqn{s}.
`scale` can be any positive number.  Defaults to `1`.}

\item{shape}{The scale parameter \eqn{\alpha}.
`shape` can be any positive number.  Defaults to `1`.}
}
\value{
A `RevWeibull` object.
}
\description{
The reversed (or negated) Weibull distribution is a special case of the
`\link{GEV}` distribution, obtained when the GEV shape parameter \eqn{\xi}
is negative.  It may be referred to as a type III extreme value
distribution.
}
\details{
We recommend reading this documentation on
  <https://alexpghayes.github.io/distributions3/>, where the math
  will render with additional detail and much greater clarity.

  In the following, let \eqn{X} be a reversed Weibull random variable with
  location parameter  `location` = \eqn{m}, scale parameter `scale` =
  \eqn{s}, and shape parameter `shape` = \eqn{\alpha}.
  An RevWeibull(\eqn{m, s, \alpha}) distribution is equivalent to a
  `\link{GEV}`(\eqn{m - s, s / \alpha, -1 / \alpha}) distribution.

  If \eqn{X} has an RevWeibull(\eqn{m, \lambda, k}) distribution then
  \eqn{m - X} has a `\link{Weibull}`(\eqn{k, \lambda}) distribution,
  that is, a Weibull distribution with shape parameter \eqn{k} and scale
  parameter \eqn{\lambda}.

  **Support**: \eqn{(-\infty, m)}.

  **Mean**: \eqn{m + s\Gamma(1 + 1/\alpha)}.

  **Median**: \eqn{m + s(\ln 2)^{1/\alpha}}{m + s(\ln 2)^(1/\alpha)}.

  **Variance**:
  \eqn{s^2 [\Gamma(1 + 2 / \alpha) - \Gamma(1 + 1 / \alpha)^2]}.

  **Probability density function (p.d.f)**:

  \deqn{f(x) = \alpha s ^ {-1} [-(x - m) / s] ^ {\alpha - 1}%
        \exp\{-[-(x - m) / s] ^ {\alpha} \}}{%
       f(x) = (\alpha / s) [-(x - m) / s] ^ (\alpha - 1)%
        exp{-[-(x - m) / s] ^ \alpha}}
  for \eqn{x < m}.  The p.d.f. is 0 for \eqn{x \geq m}{x >= m}.

  **Cumulative distribution function (c.d.f)**:

  \deqn{F(x) = \exp\{-[-(x - m) / s] ^ {\alpha} \}}{%
       F(x) = exp{-[-(x - m) / s] ^ \alpha}}
  for \eqn{x < m}.  The c.d.f. is 1 for \eqn{x \geq m}{x >= m}.
}
\examples{

set.seed(27)

X <- RevWeibull(1, 2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{Beta}()},
\code{\link{Cauchy}()},
\code{\link{ChiSquare}()},
\code{\link{Erlang}()},
\code{\link{Exponential}()},
\code{\link{FisherF}()},
\code{\link{Frechet}()},
\code{\link{GEV}()},
\code{\link{GP}()},
\code{\link{Gamma}()},
\code{\link{Gumbel}()},
\code{\link{LogNormal}()},
\code{\link{Logistic}()},
\code{\link{Normal}()},
\code{\link{StudentsT}()},
\code{\link{Tukey}()},
\code{\link{Uniform}()},
\code{\link{Weibull}()}
}
\concept{continuous distributions}
