% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernoulli.R
\name{suff_stat.Bernoulli}
\alias{suff_stat.Bernoulli}
\title{Compute the sufficient statistics for a Bernoulli distribution from data}
\usage{
\method{suff_stat}{Bernoulli}(d, x, ...)
}
\arguments{
\item{d}{A `Bernoulli` object.}

\item{x}{A vector of zeroes and ones.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the Bernoulli
  distribution:

  - `successes`: The number of successful trials (`sum(x == 1)`)
  - `failures`: The number of failed trials (`sum(x == 0)`).
}
\description{
Compute the sufficient statistics for a Bernoulli distribution from data
}
