% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{suff_stat.Binomial}
\alias{suff_stat.Binomial}
\title{Compute the sufficient statistics for the Binomial distribution from data}
\usage{
\method{suff_stat}{Binomial}(d, x, ...)
}
\arguments{
\item{d}{A `Binomial` object.}

\item{x}{A vector of zeroes and ones.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the Binomial
  distribution:

  - `successes`: The total number of successful trials.
  - `experiments`: The number of experiments run.
  - `trials`: The number of trials run per experiment.
}
\description{
Compute the sufficient statistics for the Binomial distribution from data
}
