% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{suff_stat.Gamma}
\alias{suff_stat.Gamma}
\title{Compute the sufficient statistics for a Gamma distribution from data}
\usage{
\method{suff_stat}{Gamma}(d, x, ...)
}
\arguments{
\item{d}{A `Gamma` object created by a call to [Gamma()].}

\item{x}{A vector to fit the Gamma distribution to.}

\item{...}{Unused.}
}
\value{
a `Gamma` object
}
\description{
- `sum`: The sum of the data.
  - `log_sum`: The log of the sum of the data.
  - `samples`: The number of samples in the data.
}
