% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReversedWeibull.R
\name{support.RevWeibull}
\alias{support.RevWeibull}
\title{Return the support of the RevWeibull distribution}
\usage{
\method{support}{RevWeibull}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An `RevWeibull` object created by a call to [RevWeibull()].}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the RevWeibull distribution
}
